% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_analysis.R
\name{calc_within_anova}
\alias{calc_within_anova}
\title{calc_within_anova}
\usage{
calc_within_anova(
  input_data,
  within = "subject_num",
  factors,
  dep = "RT",
  type = 3,
  generate_report = FALSE,
  report_path = NULL,
  print_desc_report = TRUE,
  debug = FALSE
)
}
\arguments{
\item{factors}{其他因子，应为一个字符向量}

\item{dep}{依赖变量，默认为 "RT"}

\item{type}{决定 Anova 的计算类别}

\item{raw_data}{原始数据，应为一个数据框}

\item{within_factor}{组内因子，默认为 "subject_num"}

\item{remove_outlier}{是否移除离群值，默认为 TRUE}
}
\value{
返回一个包含 ANOVA 结果的对象
}
\description{
计算组内方差分析
}
\examples{
# 假设我们有一个名为 df 的数据框，其中包含因子 "subject_num" 和 "condition"，以及依赖变量 "RT"
calc_within_anova(df, within_factor = "subject_num", factors = c("condition"), dep = "RT")
}
