% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(
  raw_data,
  target_col,
  filtered_by,
  resp_cor_col,
  resp_cor_value,
  debug = FALSE
)
}
\arguments{
\item{raw_data}{原始数据，应为一个数据框}

\item{target_col}{目标列，应为一个字符串，表示数据框中的一列}

\item{filtered_by}{用于过滤的列，应为一个字符向量，表示数据框中的多列}

\item{resp_cor_col}{响应相关列，应为一个字符串，表示数据框中的一列}

\item{resp_cor_value}{响应相关值，应为一个数值，表示响应相关列的值}

\item{debug}{是否开启调试模式，默认为 FALSE。如果为 TRUE，将打印被移除的离群值的信息}
}
\value{
返回一个数据框，其中已经移除了离群值
}
\description{
这个函数用于移除数据中的离群值。离群值被定义为在均值的三个标准差之外的值。
}
\examples{
# 假设我们有一个名为 df 的数据框，其中包含因子 "subject_num" 和 "condition"，以及依赖变量 "RT"
clean_data(df, target_col = "RT", filtered_by = c("subject_num", "condition"), resp_cor_col = "condition", resp_cor_value = 1, debug = TRUE)
}
