% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_analysis.R
\name{delta_plot}
\alias{delta_plot}
\title{delta_plot}
\usage{
delta_plot(
  input_data,
  subject_col,
  congruency_col,
  rt_col,
  color_by,
  facet_by,
  inc_value = "inc",
  factors_col,
  n_bins,
  remove_outlier = TRUE
)
}
\arguments{
\item{input_data}{输入数据，应为一个数据框}

\item{subject_col}{被试列名，默认为 "subject_num"}

\item{congruency_col}{一致性列名，默认为 "congruency"}

\item{rt_col}{反应时间列名，默认为 "RT"}

\item{color_by}{颜色分组列名}

\item{facet_by}{分面分组列名}

\item{inc_value}{一致性值，默认为 "inc"}

\item{factors_col}{其他因子列名，应为一个字符向量}

\item{n_bins}{分箱数量}

\item{remove_outlier}{是否移除离群值，默认为 TRUE}
}
\value{
返回一个 ggplot 对象，表示 delta 图
}
\description{
计算 Simon 效应并绘制 delta 图
}
\examples{
# 假设我们有一个名为 df 的数据框，其中包含因子 "subject_num" 和 "condition"，以及依赖变量 "RT"
delta_plot(df, subject_col = "subject_num", congruency_col = "congruency", rt_col = "RT", color_by = "condition", facet_by = "condition", factors_col = c("condition"), n_bins = 5)
}
