% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_analysis.R
\name{pc_simon_plot}
\alias{pc_simon_plot}
\title{pc_simon_plot}
\usage{
pc_simon_plot(
  input_data,
  dep_var,
  factors,
  facet_by = "condition",
  color_by = "congruency",
  ylim,
  debug = FALSE
)
}
\arguments{
\item{dep_var}{依赖变量，应为一个字符串，表示数据框中的一列}

\item{factors}{因子，应为一个字符向量，表示数据框中的多列}

\item{raw_data}{原始数据，应为一个数据框}
}
\value{
返回一个 ggplot 对象，表示绘制的条形图
}
\description{
这个函数用于绘制 Simon 任务的反应时间数据的条形图。
}
\examples{
# 假设我们有一个名为 df 的数据框，其中包含因子 "subject_num"、"condition" 和 "trial"，以及依赖变量 "RT"
pc_simon_plot(df, dep_var = "RT", factors = c("subject_num", "condition", "trial"))
}
