% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_analysis.R
\name{rt_dist_anova_and_ttest}
\alias{rt_dist_anova_and_ttest}
\title{rt_dist_anova_and_ttest}
\usage{
rt_dist_anova_and_ttest(
  simon_effect_data,
  subject_col,
  factors_col,
  facet_col,
  print_result = TRUE
)
}
\arguments{
\item{simon_effect_data}{Simon效应数据，应为一个数据框}

\item{subject_col}{被试列名}

\item{factors_col}{因子列名，应为一个字符向量}

\item{print_result}{是否打印结果，默认为 TRUE}
}
\value{
返回一个列表，包含ANOVA结果，描述统计结果和t检验结果
}
\description{
计算重复测量ANOVA和事后检验
}
\examples{
# 假设我们有一个名为 simon_effect_data 的数据框，其中包含因子 "subject_num" 和 "condition"
rt_dist_anova_and_ttest(simon_effect_data,
    subject_col = "subject_num",
    factors_col = c("condition"), print_result = TRUE
)
}
